CoDA Two Group Hypothesis 2

Performs the calculations for the hypothesis of equal covariance with different centers for two diferent samples, using Compositional Data Analysis (CoDA) principles.

CoDA_2Group_H2(comp1, comp2)

Arguments

comp1
A matrix of observations for composition 1. Entries must be non-zero and positive.

comp2
A matrix of observations for composition 2. Entries must be non-zero and positive.

Value

A tibble with the statistic (Q), degrees of freedom (nu), p-value, and null hypothesis (H0)

References

Pawlowsky-Hlahn, V., Egozcue, J.J & Tolosna-Delgado, R. (2015). Modeling and analysis of compositional data. John Wiley & Sons.

Aitchison, J. (1986). The statistical analysis of compositional data. Chapman and Hall.

Examples

data("Hongite", package = 'compositions')
data("Kongite", package = 'compositions')
CoDA_2Group_H2(Hongite,Kongite)

#> # A tibble: 1 × 4
#>       Q    nu  pval H0                                     
#>   <dbl> <dbl> <dbl> <chr>                                  
#> 1  10.7    10 0.384 equal covariance with different centers