Save ggplot

Saves/exports a ggplot object to eps.

ggplot_export(p, name = "Figure.eps", dpi = 300, w = 7, h = 4)

Arguments

p
A ggplot object

name
Name for the exported file

dpi
Resolution for the exported file

w
With in inches

h
Height in inches

Value

Exported eps file

Details

The exported "eps" file can be manipulated further for final tweaking in the desired vector software or converted to "pdf" as is. This format allows for transparency to be preserved

Examples

p = ggplot(mtcars, aes(wt, mpg)) + geom_point() + geom_smooth(method = "lm")

#> Error in ggplot(mtcars, aes(wt, mpg)): could not find function "ggplot"

ggplot_export(p)

#> Warning: unable to load shared object '/Library/Frameworks/R.framework/Resources/library/grDevices/libs//cairo.so':
#>   dlopen(/Library/Frameworks/R.framework/Resources/library/grDevices/libs//cairo.so, 6): Library not loaded: /opt/X11/lib/libcairo.2.dylib
#>   Referenced from: /Library/Frameworks/R.framework/Versions/3.6/Resources/library/grDevices/libs/cairo.so
#>   Reason: image not found

#> Warning: failed to load cairo DLL

#> Error in print(p): object 'p' not found