Barton-Choubey criterion

Calculates the Barton-Choubey criterion for the given data and for the given level of stress (unit weight and depth).

Barton_Choubey(JRC, JCS, phi.r, unit.weight, depth)

Arguments

JRC
The Joint Roughness Coefficient (1-20)

JCS
The Joint Compressive Strength in MPa

phi.r
The residual friction angle of the rock

unit.weight
The unit weight of the rock in kN/m3

depth
The depth, in meters, to which calculate the parameters

Value

A list with 3 data frames: data (normal stress and shear stress), stress level point, and parameters (c and phi). All stress values are in MPa

References

Barton, N. & Choubey, V. (1977). The shear strength of rock joints in theory and practice. Rock Mechanichs, 10: 1-54.

Examples

JRC = 10
JCS = 30
phi.r = 26
unit.weight = 18.6
depth = 40
Barton_Choubey(JRC, JCS, phi.r, unit.weight, depth)

#> $dat
#>     sig_n   tau
#> 1    0.01 0.018
#> 2    0.02 0.032
#> 3    0.03 0.044
#> 4    0.04 0.057
#> 5    0.05 0.068
#> 6    0.06 0.080
#> 7    0.07 0.091
#> 8    0.08 0.101
#> 9    0.09 0.112
#> 10   0.10 0.122
#> 11   0.11 0.133
#> 12   0.12 0.143
#> 13   0.13 0.153
#> 14   0.14 0.163
#> 15   0.15 0.173
#> 16   0.16 0.182
#> 17   0.17 0.192
#> 18   0.18 0.201
#> 19   0.19 0.211
#> 20   0.20 0.220
#> 21   0.21 0.230
#> 22   0.22 0.239
#> 23   0.23 0.248
#> 24   0.24 0.257
#> 25   0.25 0.266
#> 26   0.26 0.275
#> 27   0.27 0.284
#> 28   0.28 0.293
#> 29   0.29 0.302
#> 30   0.30 0.311
#> 31   0.31 0.319
#> 32   0.32 0.328
#> 33   0.33 0.337
#> 34   0.34 0.345
#> 35   0.35 0.354
#> 36   0.36 0.363
#> 37   0.37 0.371
#> 38   0.38 0.380
#> 39   0.39 0.388
#> 40   0.40 0.397
#> 41   0.41 0.405
#> 42   0.42 0.413
#> 43   0.43 0.422
#> 44   0.44 0.430
#> 45   0.45 0.438
#> 46   0.46 0.446
#> 47   0.47 0.455
#> 48   0.48 0.463
#> 49   0.49 0.471
#> 50   0.50 0.479
#> 51   0.51 0.487
#> 52   0.52 0.495
#> 53   0.53 0.503
#> 54   0.54 0.511
#> 55   0.55 0.520
#> 56   0.56 0.528
#> 57   0.57 0.535
#> 58   0.58 0.543
#> 59   0.59 0.551
#> 60   0.60 0.559
#> 61   0.61 0.567
#> 62   0.62 0.575
#> 63   0.63 0.583
#> 64   0.64 0.591
#> 65   0.65 0.599
#> 66   0.66 0.606
#> 67   0.67 0.614
#> 68   0.68 0.622
#> 69   0.69 0.630
#> 70   0.70 0.637
#> 71   0.71 0.645
#> 72   0.72 0.653
#> 73   0.73 0.660
#> 74   0.74 0.668
#> 75   0.75 0.676
#> 76   0.76 0.683
#> 77   0.77 0.691
#> 78   0.78 0.699
#> 79   0.79 0.706
#> 80   0.80 0.714
#> 81   0.81 0.721
#> 82   0.82 0.729
#> 83   0.83 0.736
#> 84   0.84 0.744
#> 85   0.85 0.751
#> 86   0.86 0.759
#> 87   0.87 0.766
#> 88   0.88 0.774
#> 89   0.89 0.781
#> 90   0.90 0.789
#> 91   0.91 0.796
#> 92   0.92 0.804
#> 93   0.93 0.811
#> 94   0.94 0.818
#> 95   0.95 0.826
#> 96   0.96 0.833
#> 97   0.97 0.840
#> 98   0.98 0.848
#> 99   0.99 0.855
#> 100  1.00 0.862
#> 101  1.01 0.870
#> 102  1.02 0.877
#> 103  1.03 0.884
#> 104  1.04 0.891
#> 105  1.05 0.899
#> 106  1.06 0.906
#> 107  1.07 0.913
#> 108  1.08 0.920
#> 109  1.09 0.928
#> 110  1.10 0.935
#> 111  1.11 0.942
#> 112  1.12 0.949
#> 113  1.13 0.956
#> 114  1.14 0.963
#> 115  1.15 0.971
#> 116  1.16 0.978
#> 117  1.17 0.985
#> 118  1.18 0.992
#> 119  1.19 0.999
#> 120  1.20 1.006
#> 121  1.21 1.013
#> 122  1.22 1.020
#> 123  1.23 1.027
#> 124  1.24 1.034
#> 125  1.25 1.042
#> 126  1.26 1.049
#> 127  1.27 1.056
#> 128  1.28 1.063
#> 129  1.29 1.070
#> 130  1.30 1.077
#> 131  1.31 1.084
#> 132  1.32 1.091
#> 133  1.33 1.098
#> 134  1.34 1.105
#> 135  1.35 1.112
#> 136  1.36 1.119
#> 137  1.37 1.125
#> 138  1.38 1.132
#> 139  1.39 1.139
#> 140  1.40 1.146
#> 141  1.41 1.153
#> 142  1.42 1.160
#> 143  1.43 1.167
#> 144  1.44 1.174
#> 145  1.45 1.181
#> 146  1.46 1.188
#> 147  1.47 1.195
#> 148  1.48 1.201
#> 149  1.49 1.208
#> 150  1.50 1.215
#> 151  1.51 1.222
#> 152  1.52 1.229
#> 153  1.53 1.236
#> 154  1.54 1.242
#> 155  1.55 1.249
#> 156  1.56 1.256
#> 157  1.57 1.263
#> 158  1.58 1.270
#> 159  1.59 1.276
#> 160  1.60 1.283
#> 161  1.61 1.290
#> 162  1.62 1.297
#> 163  1.63 1.304
#> 164  1.64 1.310
#> 165  1.65 1.317
#> 166  1.66 1.324
#> 167  1.67 1.330
#> 168  1.68 1.337
#> 169  1.69 1.344
#> 170  1.70 1.351
#> 171  1.71 1.357
#> 172  1.72 1.364
#> 173  1.73 1.371
#> 174  1.74 1.377
#> 175  1.75 1.384
#> 176  1.76 1.391
#> 177  1.77 1.397
#> 178  1.78 1.404
#> 179  1.79 1.411
#> 180  1.80 1.417
#> 181  1.81 1.424
#> 182  1.82 1.431
#> 183  1.83 1.437
#> 184  1.84 1.444
#> 185  1.85 1.451
#> 186  1.86 1.457
#> 187  1.87 1.464
#> 188  1.88 1.470
#> 189  1.89 1.477
#> 190  1.90 1.484
#> 191  1.91 1.490
#> 192  1.92 1.497
#> 193  1.93 1.503
#> 194  1.94 1.510
#> 195  1.95 1.516
#> 196  1.96 1.523
#> 197  1.97 1.530
#> 198  1.98 1.536
#> 199  1.99 1.543
#> 200  2.00 1.549
#> 201  2.01 1.556
#> 202  2.02 1.562
#> 203  2.03 1.569
#> 204  2.04 1.575
#> 205  2.05 1.582
#> 206  2.06 1.588
#> 207  2.07 1.595
#> 208  2.08 1.601
#> 209  2.09 1.608
#> 210  2.10 1.614
#> 211  2.11 1.621
#> 212  2.12 1.627
#> 213  2.13 1.634
#> 214  2.14 1.640
#> 215  2.15 1.647
#> 216  2.16 1.653
#> 217  2.17 1.659
#> 218  2.18 1.666
#> 219  2.19 1.672
#> 220  2.20 1.679
#> 221  2.21 1.685
#> 222  2.22 1.692
#> 223  2.23 1.698
#> 
#> $stress.level
#>       x         y
#> 1 0.744 0.6712063
#> 
#> $parameters
#>       c  phi
#> 1 0.102 37.4
#>