Moving average filter

Filters data using moving averages of diferent window sizes.

moving_avg_filter(x, y, k)

Arguments

x
A numeric vector with the position values of the data to filter (a sequence)

y
A numeric vector with the data to filter

k
A numeric vector with the widths of the windows to use for the filter (must be an odd number)

Value

A list with two data frames: one with the filtered results plus the original data, and one with the residuals

References

Swan, A. R. H. & Sandilands, M. (1995). Introduction to Geological Data Analysis. Blackwell Science.

Examples

data(nautilus)
head(nautilus)

#>   x     y
#> 1 1  4.89
#> 2 2 -4.11
#> 3 3 -4.12
#> 4 4  0.86
#> 5 5  1.86
#> 6 6 -2.14

k = c(3, 5, 11, 25)
moving_avg_filter(nautilus$x, nautilus$y, k)

#> $Filtered
#>       x     y          k_3       k_5        k_11       k_25
#> 1     1  4.89  0.390000000 -1.113333 -0.46000000 -1.3823077
#> 2     2 -4.11 -1.113333333 -0.620000 -0.70142857 -1.7978571
#> 3     3 -4.12 -2.456666667 -0.124000 -0.63375000 -1.4926667
#> 4     4  0.86 -0.466666667 -1.530000 -1.02555556 -1.2262500
#> 5     5  1.86  0.193333333 -1.138000 -1.24000000 -0.9917647
#> 6     6 -2.14 -0.810000000 -0.346000 -0.96272727 -0.9494444
#> 7     7 -2.15 -1.483333333 -1.350000 -1.87909091 -0.9121053
#> 8     8 -0.16 -2.156666667 -2.356000 -1.70454545 -0.9265000
#> 9     9 -4.16 -2.496666667 -1.566000 -1.98454545 -1.0376190
#> 10   10 -3.17 -1.840000000 -2.174000 -1.81000000 -1.1386364
#> 11   11  1.81 -2.183333333 -2.580000 -1.72727273 -1.0143478
#> 12   12 -5.19 -1.856666667 -3.188000 -1.28181818 -0.9008333
#> 13   13 -2.19 -4.860000000 -1.998000 -1.10727273 -0.8368000
#> 14   14 -7.20 -2.203333333 -1.806000 -1.11454545 -1.0440000
#> 15   15  2.78 -0.550000000 -0.216000 -0.84545455 -0.8916000
#> 16   16  2.77  2.770000000  0.176000 -0.85363636 -0.7792000
#> 17   17  2.76  1.766666667  1.568000 -1.31454545 -0.6668000
#> 18   18 -0.23  0.763333333  0.772000 -0.68636364 -0.5944000
#> 19   19 -0.24 -0.556666667 -0.434000 -0.33181818 -0.4024000
#> 20   20 -1.20 -1.566666667 -1.638000  0.38636364 -0.3300000
#> 21   21 -3.26 -2.573333333 -1.248000  0.10727273 -0.3776000
#> 22   22 -3.26 -1.600000000 -0.858000 -0.17181818 -0.2256000
#> 23   23  1.72  0.056666667 -0.478000 -0.54181818 -0.1532000
#> 24   24  1.71  1.376666667  0.116000 -0.18727273 -0.5204000
#> 25   25  0.70  0.706666667  0.708000  0.16818182 -0.3280000
#> 26   26 -0.29  0.036666667  0.102000  0.51909091 -0.2960000
#> 27   27 -0.30 -0.633333333  0.494000  0.78454545 -0.0636000
#> 28   28 -1.31  0.686666667  1.088000  0.95818182 -0.0312000
#> 29   29  3.67  2.010000000  1.678000  0.76909091 -0.1984000
#> 30   30  3.67  3.333333333  1.670000  0.49000000 -0.0460000
#> 31   31  2.66  1.996666667  1.662000 -0.24363636 -0.0140000
#> 32   32 -0.34  0.323333333  0.856000 -0.25181818 -0.0616000
#> 33   33 -1.35 -0.683333333 -0.150000 -0.35090909 -0.0312000
#> 34   34 -0.36 -1.023333333 -2.156000 -0.35818182  0.0412000
#> 35   35 -1.36 -3.030000000 -2.164000 -0.36545455  0.0736000
#> 36   36 -7.37 -3.036666667 -2.172000 -0.82727273  0.1860000
#> 37   37 -0.38 -3.046666667 -2.378000 -0.47181818  0.1784000
#> 38   38 -1.39 -1.053333333 -1.388000 -0.38909091  0.2908000
#> 39   39 -1.39  0.270000000 -0.196000 -0.39636364  0.4024000
#> 40   40  3.59  0.263333333  1.194000 -0.40363636  0.5544000
#> 41   41 -1.41  2.916666667  1.586000 -0.41181818  0.7064000
#> 42   42  6.57  1.910000000  1.578000  0.03545455  0.6188000
#> 43   43  0.57  1.903333333  0.772000  0.48181818  0.4512000
#> 44   44 -1.43 -0.433333333  0.764000  0.74636364  0.1240000
#> 45   45 -0.44 -1.106666667 -1.040000  1.19181818 -0.0036000
#> 46   46 -1.45 -1.446666667 -0.248000  1.09272727  0.0684000
#> 47   47 -2.45  0.210000000  0.342000  1.53909091  0.2204000
#> 48   48  4.53  1.200000000  1.132000  1.16818182  0.4124000
#> 49   49  1.52  3.186666667  1.922000  1.25090909  0.8044000
#> 50   50  3.51  2.510000000  3.112000  1.33363636  0.9164000
#> 51   51  2.50  3.170000000  2.704000  0.87181818  1.0284000
#> 52   52  3.50  2.830000000  2.696000  0.68272727  1.1404000
#> 53   53  2.49  2.490000000  1.890000  0.94636364  0.9332000
#> 54   54  1.48  1.150000000  0.286000  0.84727273  0.9656000
#> 55   55 -0.52 -1.520000000 -1.120000  1.02181818  0.9180000
#> 56   56 -5.52 -3.190000000 -1.528000  0.92363636  0.9500000
#> 57   57 -3.53 -2.866666667 -1.136000  0.91636364  1.1020000
#> 58   58  0.45  0.120000000 -0.344000  0.72636364  1.1740000
#> 59   59  3.44  2.443333333  1.246000  0.62818182  1.2460000
#> 60   60  3.44  3.103333333  2.436000  0.34909091  1.4776000
#> 61   61  2.43  2.763333333  2.628000  0.34181818  1.4300000
#> 62   62  2.42  2.086666667  2.222000  1.33272727  1.3428000
#> 63   63  1.41  1.746666667  1.216000  1.77818182  1.1356000
#> 64   64  1.41  0.410000000  0.610000  1.95272727  0.8484000
#> 65   65 -1.59 -0.260000000  1.202000  1.76363636  0.6812000
#> 66   66 -0.60  1.063333333  1.194000  1.48272727  0.6736000
#> 67   67  5.38  2.050000000  1.386000  1.56545455  0.7860000
#> 68   68  1.37  3.040000000  1.976000  1.64909091  1.0180000
#> 69   69  2.37  1.700000000  2.166000  1.46090909  1.2900000
#> 70   70  1.36  1.360000000  1.758000  1.18090909  1.5620000
#> 71   71  0.35  1.683333333  2.152000  0.90000000  1.5544000
#> 72   72  3.34  2.343333333  1.546000  0.89272727  1.3072000
#> 73   73  3.34  2.006666667  0.940000  0.61272727  1.0196000
#> 74   74 -0.66  0.336666667 -0.066000  0.87818182  0.7724000
#> 75   75 -1.67 -2.336666667 -0.870000  1.14272727  0.4852000
#> 76   76 -4.68 -2.343333333 -1.078000  1.13545455  0.3180000
#> 77   77 -0.68 -1.020000000 -0.088000  1.40090909  0.0304000
#> 78   78  2.30  1.970000000  1.302000  1.12090909  0.2624000
#> 79   79  4.29  3.956666667  2.494000  0.56818182  0.4544000
#> 80   80  5.28  3.616666667  3.284000  0.28727273  0.4068000
#> 81   81  1.28  3.276666667  2.876000  0.09818182  0.1996000
#> 82   82  3.27  1.603333333  1.470000  0.09090909 -0.0476000
#> 83   83  0.26  0.263333333 -0.336000 -0.09909091 -0.2544000
#> 84   84 -2.74 -2.076666667 -1.342000 -0.83363636 -0.3812000
#> 85   85 -3.75 -3.413333333 -2.948000 -0.84090909 -0.6680000
#> 86   86 -3.75 -4.086666667 -3.554000 -0.93909091 -0.7156000
#> 87   87 -4.76 -3.760000000 -4.162000 -0.67454545 -0.5232000
#> 88   88 -2.77 -4.436666667 -2.570000 -1.31818182 -0.4108000
#> 89   89 -5.78 -1.446666667 -0.980000 -1.68818182 -0.1384000
#> 90   90  4.21  0.876666667  0.810000 -1.78545455  0.0536000
#> 91   91  4.20  4.200000000  0.602000 -1.70090909  0.0460000
#> 92   92  4.19  1.526666667  0.996000 -1.70818182  0.1588000
#> 93   93 -3.81 -1.143333333 -0.608000 -1.08000000 -0.0880000
#> 94   94 -3.81 -3.810000000 -2.012000 -0.45090909 -0.1752000
#> 95   95 -3.81 -3.480000000 -3.616000  0.17818182 -0.3824000
#> 96   96 -2.82 -3.486666667 -2.424000 -0.01090909 -0.5092000
#> 97   97 -3.83 -1.500000000 -0.832000 -0.01818182 -0.6364000
#> 98   98  2.15  0.823333333  0.158000 -0.20727273 -0.6036000
#> 99   99  4.15  2.480000000  1.148000  0.78545455 -0.3316000
#> 100 100  1.14  2.473333333  2.738000  1.05090909 -0.0192000
#> 101 101  2.13  2.463333333  2.730000  1.31545455  0.2132000
#> 102 102  4.12  2.786666667  3.322000  1.39818182  0.4456000
#> 103 103  2.11  4.446666667  2.916000  1.48181818  0.2780000
#> 104 104  7.11  2.776666667  2.310000  0.74818182  0.1508000
#> 105 105 -0.89  1.773333333  1.104000  0.10454545  0.0236000
#> 106 106 -0.90 -1.233333333  0.100000  0.27818182  0.2160000
#> 107 107 -1.91 -1.906666667 -2.506000  0.36181818  0.3284000
#> 108 108 -2.91 -3.580000000 -2.914000  0.26363636  0.4804000
#> 109 109 -5.92 -3.920000000 -2.124000  0.07454545  0.3128000
#> 110 110 -2.93 -1.933333333 -1.132000 -0.57000000  0.2652000
#> 111 111  3.05  1.056666667  0.058000 -0.39636364 -0.0220000
#> 112 112  3.05  3.046666667  1.248000 -0.22272727 -0.0300000
#> 113 113  3.04  2.040000000  1.838000  0.04181818 -0.1172000
#> 114 114  0.03  1.030000000  1.432000  0.21545455 -0.2044000
#> 115 115  0.02  0.356666667  1.024000  0.75272727 -0.4516000
#> 116 116  1.02  0.683333333  0.616000  0.38181818 -0.7788000
#> 117 117  1.01  1.010000000  0.410000 -0.35181818 -1.0264000
#> 118 118  1.00  0.336666667  0.404000 -1.08636364 -1.0320833
#> 119 119 -1.00 -0.003333333 -1.202000 -1.00363636 -1.0378261
#> 120 120 -0.01 -2.673333333 -2.408000 -1.10090909 -0.9981818
#> 121 121 -7.01 -4.013333333 -3.614000 -1.10727273 -0.9071429
#> 122 122 -5.02 -5.686666667 -2.624000 -1.38727273 -0.6565000
#> 123 123 -5.03 -2.033333333 -2.830000 -2.03090909 -0.5368421
#> 124 124  3.95 -0.706666667 -1.438000 -2.03818182 -0.7361111
#> 125 125 -1.04  0.953333333 -0.846000 -2.14200000 -0.9588235
#> 126 126 -0.05 -1.050000000 -1.054000 -2.37888889 -1.2087500
#> 127 127 -2.06 -2.726666667 -1.660000 -1.80000000 -1.2913333
#> 128 128 -6.07 -2.403333333 -1.815000 -1.34000000 -1.3850000
#> 129 129  0.92 -2.575000000 -2.403333 -0.72500000 -1.5700000
#> 
#> $Residual
#>       x          k_3       k_5         k_11        k_25
#> 1     1  4.500000000  6.003333  5.350000000  6.27230769
#> 2     2 -2.996666667 -3.490000 -3.408571429 -2.31214286
#> 3     3 -1.663333333 -3.996000 -3.486250000 -2.62733333
#> 4     4  1.326666667  2.390000  1.885555556  2.08625000
#> 5     5  1.666666667  2.998000  3.100000000  2.85176471
#> 6     6 -1.330000000 -1.794000 -1.177272727 -1.19055556
#> 7     7 -0.666666667 -0.800000 -0.270909091 -1.23789474
#> 8     8  1.996666667  2.196000  1.544545455  0.76650000
#> 9     9 -1.663333333 -2.594000 -2.175454545 -3.12238095
#> 10   10 -1.330000000 -0.996000 -1.360000000 -2.03136364
#> 11   11  3.993333333  4.390000  3.537272727  2.82434783
#> 12   12 -3.333333333 -2.002000 -3.908181818 -4.28916667
#> 13   13  2.670000000 -0.192000 -1.082727273 -1.35320000
#> 14   14 -4.996666667 -5.394000 -6.085454545 -6.15600000
#> 15   15  3.330000000  2.996000  3.625454545  3.67160000
#> 16   16  0.000000000  2.594000  3.623636364  3.54920000
#> 17   17  0.993333333  1.192000  4.074545455  3.42680000
#> 18   18 -0.993333333 -1.002000  0.456363636  0.36440000
#> 19   19  0.316666667  0.194000  0.091818182  0.16240000
#> 20   20  0.366666667  0.438000 -1.586363636 -0.87000000
#> 21   21 -0.686666667 -2.012000 -3.367272727 -2.88240000
#> 22   22 -1.660000000 -2.402000 -3.088181818 -3.03440000
#> 23   23  1.663333333  2.198000  2.261818182  1.87320000
#> 24   24  0.333333333  1.594000  1.897272727  2.23040000
#> 25   25 -0.006666667 -0.008000  0.531818182  1.02800000
#> 26   26 -0.326666667 -0.392000 -0.809090909  0.00600000
#> 27   27  0.333333333 -0.794000 -1.084545455 -0.23640000
#> 28   28 -1.996666667 -2.398000 -2.268181818 -1.27880000
#> 29   29  1.660000000  1.992000  2.900909091  3.86840000
#> 30   30  0.336666667  2.000000  3.180000000  3.71600000
#> 31   31  0.663333333  0.998000  2.903636364  2.67400000
#> 32   32 -0.663333333 -1.196000 -0.088181818 -0.27840000
#> 33   33 -0.666666667 -1.200000 -0.999090909 -1.31880000
#> 34   34  0.663333333  1.796000 -0.001818182 -0.40120000
#> 35   35  1.670000000  0.804000 -0.994545455 -1.43360000
#> 36   36 -4.333333333 -5.198000 -6.542727273 -7.55600000
#> 37   37  2.666666667  1.998000  0.091818182 -0.55840000
#> 38   38 -0.336666667 -0.002000 -1.000909091 -1.68080000
#> 39   39 -1.660000000 -1.194000 -0.993636364 -1.79240000
#> 40   40  3.326666667  2.396000  3.993636364  3.03560000
#> 41   41 -4.326666667 -2.996000 -0.998181818 -2.11640000
#> 42   42  4.660000000  4.992000  6.534545455  5.95120000
#> 43   43 -1.333333333 -0.202000  0.088181818  0.11880000
#> 44   44 -0.996666667 -2.194000 -2.176363636 -1.55400000
#> 45   45  0.666666667  0.600000 -1.631818182 -0.43640000
#> 46   46 -0.003333333 -1.202000 -2.542727273 -1.51840000
#> 47   47 -2.660000000 -2.792000 -3.989090909 -2.67040000
#> 48   48  3.330000000  3.398000  3.361818182  4.11760000
#> 49   49 -1.666666667 -0.402000  0.269090909  0.71560000
#> 50   50  1.000000000  0.398000  2.176363636  2.59360000
#> 51   51 -0.670000000 -0.204000  1.628181818  1.47160000
#> 52   52  0.670000000  0.804000  2.817272727  2.35960000
#> 53   53  0.000000000  0.600000  1.543636364  1.55680000
#> 54   54  0.330000000  1.194000  0.632727273  0.51440000
#> 55   55  1.000000000  0.600000 -1.541818182 -1.43800000
#> 56   56 -2.330000000 -3.992000 -6.443636364 -6.47000000
#> 57   57 -0.663333333 -2.394000 -4.446363636 -4.63200000
#> 58   58  0.330000000  0.794000 -0.276363636 -0.72400000
#> 59   59  0.996666667  2.194000  2.811818182  2.19400000
#> 60   60  0.336666667  1.004000  3.090909091  1.96240000
#> 61   61 -0.333333333 -0.198000  2.088181818  1.00000000
#> 62   62  0.333333333  0.198000  1.087272727  1.07720000
#> 63   63 -0.336666667  0.194000 -0.368181818  0.27440000
#> 64   64  1.000000000  0.800000 -0.542727273  0.56160000
#> 65   65 -1.330000000 -2.792000 -3.353636364 -2.27120000
#> 66   66 -1.663333333 -1.794000 -2.082727273 -1.27360000
#> 67   67  3.330000000  3.994000  3.814545455  4.59400000
#> 68   68 -1.670000000 -0.606000 -0.279090909  0.35200000
#> 69   69  0.670000000  0.204000  0.909090909  1.08000000
#> 70   70  0.000000000 -0.398000  0.179090909 -0.20200000
#> 71   71 -1.333333333 -1.802000 -0.550000000 -1.20440000
#> 72   72  0.996666667  1.794000  2.447272727  2.03280000
#> 73   73  1.333333333  2.400000  2.727272727  2.32040000
#> 74   74 -0.996666667 -0.594000 -1.538181818 -1.43240000
#> 75   75  0.666666667 -0.800000 -2.812727273 -2.15520000
#> 76   76 -2.336666667 -3.602000 -5.815454545 -4.99800000
#> 77   77  0.340000000 -0.592000 -2.080909091 -0.71040000
#> 78   78  0.330000000  0.998000  1.179090909  2.03760000
#> 79   79  0.333333333  1.796000  3.721818182  3.83560000
#> 80   80  1.663333333  1.996000  4.992727273  4.87320000
#> 81   81 -1.996666667 -1.596000  1.181818182  1.08040000
#> 82   82  1.666666667  1.800000  3.179090909  3.31760000
#> 83   83 -0.003333333  0.596000  0.359090909  0.51440000
#> 84   84 -0.663333333 -1.398000 -1.906363636 -2.35880000
#> 85   85 -0.336666667 -0.802000 -2.909090909 -3.08200000
#> 86   86  0.336666667 -0.196000 -2.810909091 -3.03440000
#> 87   87 -1.000000000 -0.598000 -4.085454545 -4.23680000
#> 88   88  1.666666667 -0.200000 -1.451818182 -2.35920000
#> 89   89 -4.333333333 -4.800000 -4.091818182 -5.64160000
#> 90   90  3.333333333  3.400000  5.995454545  4.15640000
#> 91   91  0.000000000  3.598000  5.900909091  4.15400000
#> 92   92  2.663333333  3.194000  5.898181818  4.03120000
#> 93   93 -2.666666667 -3.202000 -2.730000000 -3.72200000
#> 94   94  0.000000000 -1.798000 -3.359090909 -3.63480000
#> 95   95 -0.330000000 -0.194000 -3.988181818 -3.42760000
#> 96   96  0.666666667 -0.396000 -2.809090909 -2.31080000
#> 97   97 -2.330000000 -2.998000 -3.811818182 -3.19360000
#> 98   98  1.326666667  1.992000  2.357272727  2.75360000
#> 99   99  1.670000000  3.002000  3.364545455  4.48160000
#> 100 100 -1.333333333 -1.598000  0.089090909  1.15920000
#> 101 101 -0.333333333 -0.600000  0.814545455  1.91680000
#> 102 102  1.333333333  0.798000  2.721818182  3.67440000
#> 103 103 -2.336666667 -0.806000  0.628181818  1.83200000
#> 104 104  4.333333333  4.800000  6.361818182  6.95920000
#> 105 105 -2.663333333 -1.994000 -0.994545455 -0.91360000
#> 106 106  0.333333333 -1.000000 -1.178181818 -1.11600000
#> 107 107 -0.003333333  0.596000 -2.271818182 -2.23840000
#> 108 108  0.670000000  0.004000 -3.173636364 -3.39040000
#> 109 109 -2.000000000 -3.796000 -5.994545455 -6.23280000
#> 110 110 -0.996666667 -1.798000 -2.360000000 -3.19520000
#> 111 111  1.993333333  2.992000  3.446363636  3.07200000
#> 112 112  0.003333333  1.802000  3.272727273  3.08000000
#> 113 113  1.000000000  1.202000  2.998181818  3.15720000
#> 114 114 -1.000000000 -1.402000 -0.185454545  0.23440000
#> 115 115 -0.336666667 -1.004000 -0.732727273  0.47160000
#> 116 116  0.336666667  0.404000  0.638181818  1.79880000
#> 117 117  0.000000000  0.600000  1.361818182  2.03640000
#> 118 118  0.663333333  0.596000  2.086363636  2.03208333
#> 119 119 -0.996666667  0.202000  0.003636364  0.03782609
#> 120 120  2.663333333  2.398000  1.090909091  0.98818182
#> 121 121 -2.996666667 -3.396000 -5.902727273 -6.10285714
#> 122 122  0.666666667 -2.396000 -3.632727273 -4.36350000
#> 123 123 -2.996666667 -2.200000 -2.999090909 -4.49315789
#> 124 124  4.656666667  5.388000  5.988181818  4.68611111
#> 125 125 -1.993333333 -0.194000  1.102000000 -0.08117647
#> 126 126  1.000000000  1.004000  2.328888889  1.15875000
#> 127 127  0.666666667 -0.400000 -0.260000000 -0.76866667
#> 128 128 -3.666666667 -4.255000 -4.730000000 -4.68500000
#> 129 129  3.495000000  3.323333  1.645000000  2.49000000
#>